test: break up test_distributed.py
